Add test workflow and test new()
Add the `input` and `output` params to `new()`, so that tests can
control where data is read from and written to. Neither used yet, but
use `pipe()` to create read and write file handles to test `new()`.
First tried to use file handles opened on scalars, but the GitHub
workflow runner hit an assertion failure. So `pipe()` it is.