Replace WWW::Curl::Simple with WWW::Curl::Easy
`Simple` only handles GET and POST requests, unfortunately, so skip it
and use `Easy` directly. Add a new method, `_curl`, that creates an
`Easy` with everything properly configured, including a custom user
agent, basic auth headers, other headers, the request body, and CA
validation. It returns the object and references to scalars into which
the headers and content will be populated when it executes.

Revamp the `t/demo.t` to do basic validation that the method works, then
add `t/request.t` to do live testing of its functionality. Have it rely
on an environment variable for the base URL it tests against, and load
the `go-httpbin` service for local, reliable testing. For whatever
reason it returns 100 for requests with a body while httpbingo.org
returns 200, so the test accepts either response code.