Type lines by default
Add `type_chars` param to new and teach `type` to ignore everything but
Enter keys by default, and to go back to typing all characters on a line
until the end of the line if `type_chars` is true.

Also remove unneeded modules and Apt packages and fix the name of
`type_run_psql` in the documentation.