Add `clear` option, fix multi-line `type_run_psql`
Pass a false `clear` option and the clear method will be a no-op.

Fix a bug in `type_run_psql` where it failed to print the first
character of lines 2-n.

Tweak the format of type_run_psql to use a here doc and to indent the
lines in the heredoc.