Quote URLs & HTTP data in output
Use String::ShellQuote to quote URLs, and HTTP data arguments in the
output. Update the docs to reflect other contexts in which the user is
required to ensure that shell quoting is correct.

Also add tests to cover this behavior and the behavior of the `clear`
attribute added in 7e619bc.

Increment to v0.31.0.